working more on modularizing code. Created radio buttons. Working on sorting entries by mood.
