+++ src/index.html
@@ -30,7 +30,7 @@
             <textarea name="journal" id="contents" cols="50" rows="10"></textarea>
             <br>
             <p>Mood</p>
-            <select name="mood" id="mood">
+             <select name="mood" id="mood">
                 <option value="confident">Confident</option>
                 <option value="optimistic">Optimistic</option>
                 <option value="productive">Productive</option>
@@ -41,11 +41,43 @@
                 <option value="depressed">Depressed</option>
                 <option value="uncomfortable">Uncomfortable</option>
                 <option value="worried">Nervous/Worried</option>
-            </select>
+            </select> 
             <br>
         </fieldset>
     </form>
+    <!-- <div id="radio-buttons"></div> -->
     <button type="button" id="submit-button" class="btn btn-info">Submit</button>
+
+    <legend>Filter Journal Entries by Mood</legend>
+    <input type="radio" id="radio-confident" name="radio" value="confident" checked>
+    <label for="radio-confident">Confident</label>    
+
+    <input type="radio" id="radio-optimistic" name="radio" value="optimistic">
+    <label for="radio-optimistic">Optimistic</label> 
+
+    <input type="radio" id="radio-productive" name="radio" value="productive">
+    <label for="radio-productive">Productive</label> 
+
+    <input type="radio" id="radio-successful" name="radio" value="successful">
+    <label for="radio-successful">Successful</label> 
+
+    <input type="radio" id="radio-wizard" name="radio" value="wizard">
+    <label for="radio-wizard">Like a Coding Wizard</label> 
+
+    <input type="radio" id="radio-confused" name="radio" value="confused">
+    <label for="radio-confused">Confused</label> 
+
+    <input type="radio" id="radio-discouraged" name="radio" value="discouraged">
+    <label for="radio-discouraged">Discouraged</label> 
+
+    <input type="radio" id="radio-depressed" name="radio" value="depressed">
+    <label for="radio-depressed">Depressed</label> 
+
+    <input type="radio" id="radio-uncomfortable" name="radio" value="uncomfortable">
+    <label for="radio-uncomfortable">Uncomfortable</label> 
+
+    <input type="radio" id="radio-nervous" name="radio" value="nervous">
+    <label for="radio-nervous">Nervous/Worried</label> 
 
     <article id="journal-entries"></article>
