+++ src/scripts/journal.js
@@ -1,6 +1,6 @@
 import API from "./data.js"
+import journalSenderToApi from "./journalSenderToApi.js"
 import entriesOnDom from "./entriesDOM.js"
-import eventListener from "./eventListeners.js"
 // import eventListener from "./eventListener.js"
 //RESPONSIBILITY: INITIAL VIEW
 
@@ -18,23 +18,53 @@ import eventListener from "./eventListeners.js"
 // Invoke the render function, and the argument is the array of objects in the data.js file. 
 
 API.getJournalEntries()
-//FROM DATA.JS
-// .then(parsedEntries => {
-//     parsedEntries.forEach(entry => {
-//         entriesOnDom.renderJournalEntry(entry)
-//     })
-// })
+    //FROM DATA.JS
+    .then(parsedEntries => {
+        parsedEntries.forEach(entry => {
+            entriesOnDom.renderJournalEntry(entry)
+        })
+    })
 //at some point this will be the factory function?
-
-
 const submitButtonEventListener = document.querySelector("#submit-button")
 console.log(submitButtonEventListener)
-submitButtonEventListener.addEventListener("click", () => eventListener.sendJournalEntryToApi())
+submitButtonEventListener.addEventListener("click", () => journalSenderToApi.sendJournalEntryToApi())
+
+// const dropdownValues = document.querySelector("option")
+// console.log(dropdownValues.value)
 
 
+// const dropdownValuesToRadioButtons = (dropdownStuff) => {
+//     return `
+// <input type="radio" id="radio-${dropdownStuff}" name="radio" value="radio-${dropdownStuff}">
+// <label for="radio-${dropdownValues}">${dropdownStuff}</label>  
+// `}
 
-//Grab the boxes
-//Grab the info from the boxes
-//quality control on info
-//put them in a factory function
-//send them to the server
+// const renderRadioButtonToDom = () => {
+//  const divForRadioButtons = document.querySelector("#radio-buttons")
+//  console.log(divForRadioButtons)
+//  divForRadioButtons.innerHTML = dropdownValuesToRadioButtons("Confident")
+// }
+
+// renderRadioButtonToDom()
+
+const radioButtons = document.getElementsByName("radio")
+console.log("RADIO", radioButtons)
+radioButtons.forEach(button => {
+    button.addEventListener("click", event => {
+        let mood = event.target.value
+        console.log("MOOD", mood)
+        API.getJournalEntries()
+            .then(parsedEntries => {
+                // console.log(parsedEntries)
+                const arrayThatMatchesMood = parsedEntries.filter(entry => mood === entry.mood)
+                arrayThatMatchesMood.forEach(objInArray => {
+                    entriesOnDom.renderJournalEntry(objInArray)
+                })
+
+            })
+    })
+})
+
+
+// for(let i = 0; i <radioButtons.length; i++) {
+//     radioButtons[i].addEventListener
